Restructure evaluator hierarchy
