+++ loss_landscapes/sl/landscapes.py
@@ -5,7 +5,7 @@ Functions for approximating landscapes in one and two dimensions.
 import copy
 import numpy as np
 from loss_landscapes.common.model_interface.model_agnostic_factories import wrap_model, rand_u_like
-from loss_landscapes.common.evaluators.evaluator import Evaluator
+from loss_landscapes.common.evaluators.evaluators import Evaluator
 
 
 def point(model, evaluator: Evaluator) -> tuple:
